requirements.txt
celery==5.6.0
docling==2.65.0
docling_core==2.57.0
fastapi==0.128.0
jose==1.0.0
langchain==1.2.7
langchain_classic==1.0.1
langchain_community==0.4.1
langchain_core==1.2.7
langchain_milvus==0.3.3
langchain_ollama==1.0.1
langchain_openai==1.1.7
langchain_text_splitters==1.1.0
langgraph==1.0.7
passlib==1.7.4
psycopg==3.3.2
psycopg_pool==3.3.0
pydantic==2.12.5
pydantic_settings==2.12.0
pymilvus==2.6.5
python-dotenv==1.2.1
python_jose==3.5.0
uvicorn==0.40.0
python-multipart==0.0.21
langgraph-checkpoint-postgres==3.0.2
email-validator==2.3.0
bcrypt==4.0.1
redis==7.1.0
easyocr==1.7.2